#13 revised tests, testspictures and added examples
