+++ uesgraphs/__init__.py
@@ -1,5 +1,3 @@
 from uesgraphs.uesgraph import UESGraph
 
 from uesgraphs.visuals import Visuals
-
-from uesgraphs.examples.example_uesgraph import *
