+++ uesgraphs/examples/e1_example_readme.py
@@ -1,0 +1,80 @@
+"""This module includes tests and demonstrations of uesgraphs"""
+
+import os
+from shapely.geometry import Point
+import uesgraphs as ug
+
+import matplotlib.pyplot as plt
+
+
+def main():
+
+    example_readme()
+
+
+def example_readme():
+    """The example given in the README.md file
+    """
+    workspace = make_workspace('readme')
+
+    graph = ug.UESGraph()
+
+    supply = graph.add_building(
+        name='Supply',
+        position=Point(0, 10),
+        is_supply_heating=True,
+    )
+    demand = graph.add_building(
+        name='Building 1',
+        position=Point(50, 15),
+    )
+    heating_node = graph.add_network_node(
+        network_type='heating',
+        position=Point(30, 5),
+    )
+
+    graph.add_edge(supply, heating_node)
+    graph.add_edge(heating_node, demand)
+
+    save_as = os.path.join(workspace, 'graph.png')
+    vis = ug.Visuals(graph)
+    vis.show_network(
+        save_as=save_as,
+        scaling_factor=30,
+    )
+
+
+def make_workspace(name_workspace=None):
+    """Creates a local workspace with given name
+
+    If no name is given, the general workspace directory will be used
+
+    Parameters
+    ----------
+    name_workspace : str
+        Name of the local workspace to be created
+
+    Returns
+    -------
+    workspace : str
+        Full path to the new workspace
+    """
+    this_dir = os.path.dirname(__file__)
+    ues_dir = os.path.dirname(os.path.dirname(this_dir))
+    workspace = os.path.join(ues_dir, 'workspace')
+    if not os.path.exists(workspace):
+        os.mkdir(workspace)
+
+    if name_workspace is not None:
+        workspace = os.path.join(workspace, name_workspace)
+        if not os.path.exists(workspace):
+            os.mkdir(workspace)
+
+    return workspace
+
+
+# Main function
+if __name__ == '__main__':
+    print("*** Generating example readme ***")
+    main()
+    print("*** Done ***")
